Add serde and serde_json for json data export
This is important so that data isn't just being printed to console. We can not export it to a file and do what we want with it. An improvement would also to be to dump to CSV because that's a logical thing to do however, I couldn't get it working correctly. Will explore again later. I believe I got a fix now with the `skip` on the problematic fields.